refactor: GNB 일부 리팩토링, 불필요한 타입 제거
