+++ client/src/components/common/Gnb/index.tsx
@@ -1,10 +1,9 @@
 import React, { Dispatch, useEffect } from 'react';
 import styled, { css, keyframes } from 'styled-components';
-import { Link, useHistory } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 import { useRecoilValue, useRecoilState, useResetRecoilState } from 'recoil';
 
 import {
-  modalStateStore,
   rightModalStates,
   solvedProblemState,
   userDataStates,
@@ -12,36 +11,33 @@ import {
   alarmState,
   usersocketStates,
   themeState,
-  commonState,
-  currentPageStates
+  commonState
 } from 'recoil/store';
 import fetchApi from 'api/fetch';
-import { FlexProps, TabProps, RightModalProps } from 'types/GNB';
-import {
-  GnbHome,
-  GnbGroup,
-  GnbHomeActive,
-  GnbGroupActive,
-  GnbMessage,
-  GnbMessageActive,
-  GnbAlarm,
-  GnbAlarmActive,
-  GnbLogout
-} from 'images/icons';
-import { mainLogo } from 'images';
-import useResetProfile from 'hooks/useResetProfile';
+import { RightModalProps } from 'types/GNB';
+import { GnbMessage, GnbAlarm, GnbLogout } from 'images/icons';
 
-import { ProfilePhoto } from 'components/common';
-import UserSearchModal from './UserSearchModal';
-import UserSearchBar from './UserSearchBar';
-import { Page } from 'types/common';
+import GnbLeftItems from './GnbLeftItems';
+import GnbCenterItems from './GnbCenterItems';
+import ProfileCard from './ProfileCard';
 
-const MainLogo = styled(Link)`
-  width: 40px;
-  height: 40px;
-  display: block;
-  background-image: url(${mainLogo});
-  background-size: 40px 40px;
+const turnDark = keyframes`
+  0%{
+    left: -2px;
+  }
+  100%{
+    left: 22px;
+    background-color: #8a8c91;
+  }
+`;
+const turnLight = keyframes`
+  0%{
+    left: 22px;
+  }
+  100%{
+    left: -2px;
+    background-color: #87d474;
+  }
 `;
 
 const GnbContainer = styled.div<{ commonState: boolean }>`
@@ -65,107 +61,58 @@ const GnbContainer = styled.div<{ commonState: boolean }>`
   }
 `;
 
-const FlexWrap = styled.div<FlexProps>`
+const FlexBox = styled.div`
   display: flex;
   align-items: center;
 
   & > *:not(:first-child) {
     margin-left: 8px;
   }
-
-  ${({ center }) =>
-    center &&
-    css`
-      position: absolute;
-      left: 50%;
-      top: 50%;
-      transform: translate(-50%, -50%);
-
-      @media screen and (max-width: 852px) {
-        margin-left: 20px;
-      }
-    `}
 `;
 
-const GnbTab = styled.div<TabProps>`
-  width: 112px;
-  height: 48px;
+const ToggleBtnWrap = styled.div`
+  width: 50px;
+  height: 25px;
+
   display: flex;
   justify-content: center;
   align-items: center;
-  transition: background-color border-radius 0.1s ease-in;
 
   &:hover {
-    background-color: ${(props) => props.theme.lightgray};
-    border-radius: 8px;
-  }
-
-  &:active {
-    background-color: ${(props) => props.theme.gray};
-  }
-
-  svg path {
-    ${({ current }) =>
-      current
-        ? css`
-            fill: ${(props) => props.theme.green};
-          `
-        : css`
-            fill: ${(props) => props.theme.darkgray};
-          `}
-  }
-
-  @media screen and (max-width: 920px) {
-    width: 60px;
-    height: 48px;
+    cursor: pointer;
   }
 `;
 
-const ProfileWrap = styled.div`
-  height: 36px;
-  display: flex;
-  align-items: center;
-  padding-left: 4px;
-  padding-right: 12px;
+const ToggleBar = styled.div`
+  position: relative;
+  width: 80%;
+  height: 30%;
 
-  &:hover {
-    background-color: ${(props) => props.theme.lightgray};
-    border-radius: 24px;
-  }
-
-  &:active {
-    background-color: ${(props) => props.theme.gray};
-  }
-
-  p {
-    color: ${(props) => props.theme.black};
-    margin-left: 8px;
-    font-size: 1rem;
-    font-weight: bold;
-  }
-
-  @media screen and (max-width: 920px) {
-    padding: 0px;
-    p {
-      display: none;
-      margin-left: 0px;
-    }
-    img {
-      width: 36px;
-      height: 36px;
-
-      &:hover {
-        filter: brightness(90%);
-      }
-
-      &:active {
-        filter: brightness(80%);
-      }
-    }
-  }
+  border-radius: 20px;
+  background-color: ${(props) => props.theme.gray};
 `;
 
-const IconWrap = styled.div<TabProps>`
+const ToggleBtn = styled.div<{ themeState: string }>`
+  position: absolute;
+  top: -6px;
+  left: ${(props) => (props.themeState === 'dark' ? '22px' : '-2px')};
+  width: 20px;
+  height: 20px;
+
+  border-radius: 50%;
+  background-color: ${(props) =>
+    props.themeState === 'dark' ? props.theme.darkgray : props.theme.green};
+  animation: ${(props) =>
+    props.themeState === 'dark'
+      ? css`
+          ${turnDark} ease 0.5s
+        `
+      : css`
+          ${turnLight} ease 0.5s
+        `};
+`;
+
+const SideButton = styled.div<{ current?: boolean }>`
   width: 40px;
   height: 40px;
   background: ${(props) => props.theme.lightgray};
@@ -225,69 +172,7 @@ const AlarmBadge = styled.div`
   font-size: 8px;
 `;
 
-const ToggleBtnWrap = styled.div`
-  width: 50px;
-  height: 25px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const ToggleBar = styled.div`
-  position: relative;
-  width: 80%;
-  height: 30%;
-
-  border-radius: 20px;
-  background-color: ${(props) => props.theme.gray};
-`;
-
-const turnDark = keyframes`
-  0%{
-    left: -2px;
-  }
-  100%{
-    left: 22px;
-    background-color: #8a8c91;
-  }
-`;
-const turnLight = keyframes`
-  0%{
-    left: 22px;
-  }
-  100%{
-    left: -2px;
-    background-color: #87d474;
-  }
-`;
-
-const ToggleBtn = styled.div<{ themeState: string }>`
-  position: absolute;
-  top: -6px;
-  left: ${(props) => (props.themeState === 'dark' ? '22px' : '-2px')};
-  width: 20px;
-  height: 20px;
-
-  border-radius: 50%;
-  background-color: ${(props) =>
-    props.themeState === 'dark' ? props.theme.darkgray : props.theme.green};
-  animation: ${(props) =>
-    props.themeState === 'dark'
-      ? css`
-          ${turnDark} ease 0.5s
-        `
-      : css`
-          ${turnLight} ease 0.5s
-        `};
-`;
-
-const Gnb = ({ type }: { type?: string }) => {
-  const modalState = useRecoilValue(modalStateStore);
+const Gnb = () => {
   const [userdata, setUserdata] = useRecoilState(userDataStates);
   const resetSolvedProblemState = useResetRecoilState(solvedProblemState);
   const [rightModalState, setRightModalState] =
@@ -298,12 +183,6 @@ const Gnb = ({ type }: { type?: string }) => {
   const socket = useRecoilValue(usersocketStates);
   const [commonDisplay, setCommonDispaly] = useRecoilState(commonState);
   const history = useHistory();
-  const resetProfile = useResetProfile();
-  const currentPage = useRecoilValue(currentPageStates);
-
-  const photoClickHandler = (e: React.MouseEvent) => {
-    resetProfile(userdata.name);
-  };
 
   const themeToggleHandler = (e: React.MouseEvent) => {
     setTheme(theme === 'light' ? 'dark' : 'light');
@@ -322,40 +201,16 @@ const Gnb = ({ type }: { type?: string }) => {
 
   return (
     <GnbContainer commonState={commonDisplay} className="no-drag">
-      <FlexWrap>
-        <MainLogo to="/home" />
-        <UserSearchBar />
-        {modalState.searchUser && <UserSearchModal />}
-      </FlexWrap>
-      <FlexWrap center>
-        <Link to="/home">
-          <GnbTab current={currentPage === Page.HOME}>
-            {currentPage === Page.HOME ? <GnbHomeActive /> : <GnbHome />}
-          </GnbTab>
-        </Link>
-        <Link to="/groupselect">
-          <GnbTab current={currentPage === Page.GROUP_SELECT}>
-            {currentPage === Page.GROUP_SELECT ? (
-              <GnbGroupActive />
-            ) : (
-              <GnbGroup />
-            )}
-          </GnbTab>
-        </Link>
-      </FlexWrap>
-      <FlexWrap>
+      <GnbLeftItems />
+      <GnbCenterItems />
+      <FlexBox>
         <ToggleBtnWrap onClick={themeToggleHandler}>
           <ToggleBar>
             <ToggleBtn themeState={theme} />
           </ToggleBar>
         </ToggleBtnWrap>
-        <Link to={`/profile/${userdata.name}`} onClick={photoClickHandler}>
-          <ProfileWrap>
-            <ProfilePhoto userName={userdata.name} size="28px" />
-            <p>{userdata.name}</p>
-          </ProfileWrap>
-        </Link>
-        <IconWrap
+        <ProfileCard name={userdata.name} />
+        <SideButton
           current={rightModalState.messageFlag === true}
           onClick={() => {
             ChangeFlag(rightModalState, setRightModalState, 'messageFlag');
@@ -365,9 +220,9 @@ const Gnb = ({ type }: { type?: string }) => {
             });
           }}
         >
-          {type === 'message' ? <GnbMessageActive /> : <GnbMessage />}
-        </IconWrap>
-        <IconWrap
+          <GnbMessage />
+        </SideButton>
+        <SideButton
           current={rightModalState.alarmFlag === true}
           onClick={() => {
             ChangeFlag(rightModalState, setRightModalState, 'alarmFlag');
@@ -375,8 +230,8 @@ const Gnb = ({ type }: { type?: string }) => {
             socket.emit('make alarms check', { receiver: userdata.name });
           }}
         >
-          {type === 'alarm' ? <GnbAlarmActive /> : <GnbAlarm />}
-        </IconWrap>
+          <GnbAlarm />
+        </SideButton>
         <AlarmBadge
           onClick={() => {
             ChangeFlag(rightModalState, setRightModalState, 'alarmFlag');
@@ -386,7 +241,7 @@ const Gnb = ({ type }: { type?: string }) => {
         >
           {alarmNum ? alarmNum : null}
         </AlarmBadge>
-        <IconWrap
+        <SideButton
           onClick={async () => {
             ChangeFlag(rightModalState, setRightModalState, '');
             await fetchApi.logout();
@@ -405,8 +260,8 @@ const Gnb = ({ type }: { type?: string }) => {
           }}
         >
           {<GnbLogout />}
-        </IconWrap>
-      </FlexWrap>
+        </SideButton>
+      </FlexBox>
     </GnbContainer>
   );
 };
